chore(refactor): Prepare computeAccessibleDescription
